Added SharedPreferences and default initialization mechanism.
The existing android classes for this (with the exception
of SharedPreferences itself) seem like overkill for what
we want to do and in any case do not seem to work reliably.
I have therefore implemented a compromise solution where the
SharedPreferences for the application are initialized from a
file named defaults.xml which contains appropriate key-value
pairs. We can therefore produce versions of the application
for different OTM instances by replacing this file.

Additional settings to be added as and when we need them.