Add model container and handler
As an extention to the model and rest handler pattern we're using, this
modelContainer and assocaited handler will handle responses where a
JSONArray is returned instead of a JSONObject.  An instance of a class
which extends model container can access the underlying model instances.

This includes an implementation for species.