Implemented tiled map with border.
This currently consists of nine tiles, each the size
of the screen. Initially, the center one is displayed.
Panning works smoothly within the nine-tiles since the
images are already loaded.

Next step is to re-center grid and load new tiles as
the user pans towards the limits of the grid.