Add sections to species list adapter
When rendering a row of the list, check to see if the first letter
of the name has already been included in a section, and if not, display
that cell's section text.

This also caches the state of the section, because scrolling or filter
will cause the list to be rerendered.  Depending on the direction of the
scroll, either the first or last element will be checked first.