Screen-size and number of tiles made more generic.
Number of tiles for X and Y axes are now read from
shared-preferences.

Screens-size is acquired from device (NOTE: this currently
assumes that the map is the only thing on screen.
Getting height and width from a MapView is problematic
but will probably have to be addressed sooner rather
than later.