Switched to map-polling to keep tiles and map sync'd.
Zooming (i.e. either scaling or reloading tiles at
a new zoom level) was working but the tiles kept
getting out of sync with the map. This was happening
to a lesser extent with panning anyway. I've therefore
changed things around so that events are propagated
straight through to the underlying OTMMapView with
WMSTileRaster otherwise ignoring them. WMSTileRaster
does, however, repeatedly poll the map in order to
keep itself synchronized.

A nice corollary of this is that we don't have to
mess about dealing with off-center pinch-zoom gestures
that would otherwise pan and zoom at the same time.