Handling map click event and showing an icon on returned plot
Modified the Geometry class to use doubles instead of longs so that fractional
degrees are possible. Also added E6 converter methods so we don't have to be
doing that all over. Added a PlotContainer class to hold the list of plots that
come back from the request generator for nearby plots. Added an up event and
a check to see if this was a previous move event. Added code to make a call out
to the rest api, getting the nearest plot for a touch event and displaying a
placeholder image at that plot's location. Dataa containser now throw internal
JSON exceptions.