Panning bug-fix for absolute coordinates.
Tiles were originally processed based upon a relative
position to the screen-origin. This ceased being the
case when WMSTileRaster was switched over to polling (for better
synchronization with the MapView. This caused a problem
whereby tile-loading initiated by panning past the available
tile-boundary would cause a cascade of load-requests.

Removed code that assumed relative pan-offsets and added
code that computes the relative offset and then compares
against previous.,