Renamed initZoomPolling to initPanPolling.
Originally, this method started a thread to monitor
the zoom-level of the MapView so that WMSTileRaster
could zoom itself accordingly. This didn't work well
so it was moved out to an event-based approach (by
subclassing MapView). During this process, it was
found that tile/map synchronization was much better
during panning (no 'tile-trailing') if the coordinates
of the MapView were polled for. This is now the thread's
sole job hence the rename of its initialization-method.