Switched to mockito and modified manifest to support tests.
The annotation-processor used by AndroidMock was
preventing ant-based building of and running of
the unit-tests so I've switched over to mockito.

I've also made Google Maps optional for the application
as it isn't necessary for the tests to run and requiring
it prevents deployment to androidx86 in VirtualBox since
that image doesn't have the Google Maps library.