Species selector now returns full species JSON
Previously the ID was returned, which would need to be loaded from
the manager.  It can just serialize the object to it's underlaying
JSON and pass it back with the return message.  This prevents having
to use the FilterManager on an additional process when it's used from
an activity that has a map (which will always be on it's own process)