Prevented display of tiles during zooming.
The tiles do not scale with the map so showing
them produces unpleasant visual effects. I've
added a drawTiles boolean that allows drawing
to be switched on and off. If the app detects
a multi-touch gesture (which must be zooming
since we don't support any others), it will
switch drawing off.

I also attempted to switch polling (to keep map
and tile-layer in sync) off when there is no
interaction with the map. This was only partially
successful since tapping on a tree causes the
map to center on its location - unfortunately
this is done asynchronously and in a different
class so polling cannot be switched off after
the animation is complete; I'll attempt to revisit
this at some point.