Fixed the map popup tree view.
The popup was trying to use the tree's species name,
but was bailing out entirely when that routine threw an exception.
I wrapped the species name code in its own
try-catch block, and set a default name when the data is not found.
It's a little weird, because this itself is in a try/catch handling
the same JSON exception.  I think it is correct though, because we want
to handle the specific "no species found" state gracefully, and then
catch everything else.

At any rate, this fixes the tree image thumbnail in that view (because
we are no longer bailing out of the code block before we get to the
thumbnail code.)

This closes the loop on edit tree photo.