Make sure shared preferences refresh
I'm not sure why this was written the way it was, but the result
was that you have to completely uninstall and reinstall the application
for defaults changes.  I'm not sure if loading shared prefs was percieved
as slow, or if we were going to store stuff in there that the user was going
to set.  Since neither of those seem the case, I am removing the 'first-run' check
from this code and refreshing the shared prefs from resources on every run.