Added another response handler so that bad logins properly dismiss the "please wait - logging in" message.
This is a fairly unsatisfying fix, and I think that is mostly due
to the poorly designed async http library.  I hypothesize that what was
happening was that on bad credentials the server was responding with no
data but a header of FORBIDDEN.  The callbacks we were overriding before
this fix did not handle that case.