Style cleanups - Fixed margins and added borders between fields
This style cleanup required changing how fields are rendered slightly
Without passing in a View's parent when it is being inflated, the root View's
layout parameters don't take affect, which made styling the tree detail page
properly impossible.

The parent ViewGroups are now properly passed down to everywhere we inflate
layouts. (Though note we pass false for the 'addToRoot' parameter, because we
do that explicitly later).