Prevent app crashes from infinite loops in the DiameterField event handler
Changes to diameter text field were triggering the event handler for the
circumference field, which could then change the text for the diameter
field and trigger another event (ad infinitum). We attempted to prevent
this by not setting the text if the value was not different from what
was already present, but exceptionally large or small values do not
"round trip" through the diameter/circumference conversion perfectly and
triggered an infinite amount of text changes, causing the app to crash.

By tracking whether we are currently setting the text and not handling any
text changes while we are doing so, we avoid this problem.

Connects to #247