Fix map NPE
Rollbar registered the following error:
```
java.lang.NullPointerException: Attempt to invoke virtual method
'void com.google.android.gms.maps.GoogleMap.moveCamera(
    com.google.android.gms.maps.CameraUpdate)'
on a null object reference

at org.azavea.otm.ui.MainMapFragment.lambda$null$3
    (MainMapFragment.java:488)
```

It's not clear whether the null refers to the `CameraUpdate`
from the `CameraUpdateFactory` or to `mMap`, but I'm guessing `mMap`.

If that's the case, it's probably a race condition,
which explains why I can't reproduce it.

Following instructions in
[MapView](https://developers.google.com/android/reference/com/google/android/gms/maps/MapView)
to call `mapView` "on" methods required for the fragment lifecycle.

The instructions say to call `mapView.onCreate()` from the fragment
`onCreate`, but we're dependent on the view, so that has to be in
the fragment `onCreateView`. I don't know if that could be a source
of race condition.

Since we can't follow the lifecycle instructions for `onCreate`,
put up a defensive null check.